src/test/javascript/spec/app/entities/agreement/agreement.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import * as moment from 'moment';
import { DATE_TIME_FORMAT } from 'app/shared/constants/input.constants';
import { AgreementService } from 'app/entities/agreement/agreement.service';
import { IAgreement, Agreement } from 'app/shared/model/agreement.model';

describe('Service Tests', () => {
    describe('Agreement Service', () => {
        let injector: TestBed;
        let service: AgreementService;
        let httpMock: HttpTestingController;
        let elemDefault: IAgreement;
        let currentDate: moment.Moment;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(AgreementService);
            httpMock = injector.get(HttpTestingController);
            currentDate = moment();

            elemDefault = new Agreement(0, 'AAAAAAA', currentDate, currentDate, 'AAAAAAA', 'AAAAAAA');
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign(
                    {
                        agreementStartDate: currentDate.format(DATE_TIME_FORMAT),
                        agreementEndDate: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a Agreement', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0,
                        agreementStartDate: currentDate.format(DATE_TIME_FORMAT),
                        agreementEndDate: currentDate.format(DATE_TIME_FORMAT)
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        agreementStartDate: currentDate,
                        agreementEndDate: currentDate
                    },
                    returnedFromService
                );
                service
                    .create(new Agreement(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a Agreement', async () => {
                const returnedFromService = Object.assign(
                    {
                        agreementName: 'BBBBBB',
                        agreementStartDate: currentDate.format(DATE_TIME_FORMAT),
                        agreementEndDate: currentDate.format(DATE_TIME_FORMAT),
                        sourceId: 'BBBBBB',
                        source: 'BBBBBB'
                    },
                    elemDefault
                );

                const expected = Object.assign(
                    {
                        agreementStartDate: currentDate,
                        agreementEndDate: currentDate
                    },
                    returnedFromService
                );
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of Agreement', async () => {
                const returnedFromService = Object.assign(
                    {
                        agreementName: 'BBBBBB',
                        agreementStartDate: currentDate.format(DATE_TIME_FORMAT),
                        agreementEndDate: currentDate.format(DATE_TIME_FORMAT),
                        sourceId: 'BBBBBB',
                        source: 'BBBBBB'
                    },
                    elemDefault
                );
                const expected = Object.assign(
                    {
                        agreementStartDate: currentDate,
                        agreementEndDate: currentDate
                    },
                    returnedFromService
                );
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a Agreement', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
